requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
weaviate-client==4.17.0
openai==2.0.0
PyPDF2==3.0.1
python-dotenv==1.1.1
